feat: added pills for funds page filtering
