+++ src/components/frame/v5/pages/FundsPage/partials/Filter/Filter.tsx
@@ -1,18 +1,11 @@
-import { CaretDown, MagnifyingGlass } from '@phosphor-icons/react';
-import { AnimatePresence, motion } from 'framer-motion';
-import cloneDeep from 'lodash/cloneDeep';
-import get from 'lodash/get';
-import set from 'lodash/set';
+import { MagnifyingGlass } from '@phosphor-icons/react';
 import React, { useCallback, useState } from 'react';
 import { usePopperTooltip } from 'react-popper-tooltip';
 
-import { accordionAnimation } from '~constants/accordionAnimation.ts';
 import { useMobile } from '~hooks/index.ts';
 import useToggle from '~hooks/useToggle/index.ts';
 import { formatMessage } from '~utils/yup/tests/helpers.ts';
-import Checkbox from '~v5/common/Checkbox/index.ts';
 import SearchInputMobile from '~v5/common/Filter/partials/SearchInput/SearchInput.tsx';
-import AccordionItem from '~v5/shared/Accordion/partials/AccordionItem/index.ts';
 import Button from '~v5/shared/Button/Button.tsx';
 import FilterButton from '~v5/shared/Filter/FilterButton.tsx';
 import Modal from '~v5/shared/Modal/index.ts';
@@ -20,192 +13,10 @@ import PopoverBase from '~v5/shared/PopoverBase/index.ts';
 import SearchInputDesktop from '~v5/shared/SearchSelect/partials/SearchInput/index.ts';
 import Header from '~v5/shared/SubNavigationItem/partials/Header.tsx';
 
-import {
-  type RootFilterProps,
-  type FilterProps,
-  type FilterValue,
-  type NestedFilterProps,
-} from './types.ts';
+import FilterItem from './FilterItem.tsx';
+import { type FilterProps, type FilterValue } from './types.ts';
 
 const displayName = 'v5.pages.FundsPage.partials.Filter';
-
-function NestedFilterItem<TValue extends FilterValue, TLevel extends number>({
-  label,
-  path,
-  value,
-  onChange,
-  items: nestedItems,
-}: NestedFilterProps<TValue, TLevel>) {
-  const {
-    getTooltipProps: getNestedTooltipProps,
-    setTooltipRef: setNestedTooltipRef,
-    setTriggerRef: setNestedTriggerRef,
-  } = usePopperTooltip({
-    delayShow: 200,
-    delayHide: 200,
-    placement: 'left-start',
-    interactive: true,
-  });
-  const isChecked = !!get(value, path);
-  const isMobile = useMobile();
-  const NestedItems = nestedItems?.map(({ ...nestedItem }) => (
-    // @ts-ignore
-    <NestedFilterItem
-      key={nestedItem.name}
-      onChange={onChange}
-      path={`${path}.${nestedItem.name}`}
-      value={value}
-      {...nestedItem}
-    />
-  ));
-
-  return (
-    <>
-      {isMobile ? (
-        <>
-          <Checkbox
-            isChecked={isChecked}
-            onChange={(e) => {
-              onChange(set(cloneDeep(value), path, e.target.checked));
-            }}
-            classNames="subnav-button px-0 sm:px-3.5"
-          >
-            {label}
-          </Checkbox>
-          {!!nestedItems?.length && (
-            <AnimatePresence>
-              {isChecked && (
-                <motion.div
-                  key="accordion-content"
-                  initial="hidden"
-                  animate="visible"
-                  exit="hidden"
-                  variants={accordionAnimation}
-                  transition={{ duration: 0.4, ease: 'easeOut' }}
-                  className="overflow-hidden pl-4"
-                >
-                  {NestedItems}
-                </motion.div>
-              )}
-            </AnimatePresence>
-          )}
-        </>
-      ) : (
-        <>
-          <div ref={setNestedTriggerRef}>
-            <Checkbox
-              isChecked={isChecked}
-              onChange={(e) => {
-                onChange(set(cloneDeep(value), path, e.target.checked));
-              }}
-              classNames="subnav-button px-0 sm:px-3.5"
-            >
-              {label}
-            </Checkbox>
-          </div>
-          {!!nestedItems?.length && isChecked && (
-            <PopoverBase
-              setTooltipRef={setNestedTooltipRef}
-              tooltipProps={getNestedTooltipProps}
-              withTooltipStyles={false}
-              cardProps={{
-                rounded: 's',
-                hasShadow: true,
-                className: 'py-6 px-2',
-              }}
-              classNames="w-full sm:max-w-[13.25rem] mr-2"
-            >
-              {NestedItems}
-            </PopoverBase>
-          )}
-        </>
-      )}
-    </>
-  );
-}
-
-function RootFilter<TValue extends FilterValue>({
-  items,
-  label,
-  icon: Icon,
-  onChange,
-  path,
-  value,
-  title,
-}: RootFilterProps<TValue>) {
-  const { getTooltipProps, setTooltipRef, setTriggerRef, visible } =
-    usePopperTooltip({
-      delayShow: 200,
-      delayHide: 200,
-      placement: 'left-start',
-      interactive: true,
-    });
-  const isMobile = useMobile();
-  const [isAccordionOpen, { toggle: toggleAccordion }] = useToggle({
-    defaultToggleState: true,
-  });
-  const RootItems = items.map(
-    ({ label: nestedFilterItemLabel, name, items: nestedFilterItems }) => (
-      <NestedFilterItem<TValue, 2>
-        key={name.toString()}
-        label={nestedFilterItemLabel}
-        name={name}
-        items={nestedFilterItems}
-        onChange={onChange}
-        value={value}
-        path={`${path}.${name.toString()}`}
-      />
-    ),
-  );
-
-  return (
-    <>
-      {isMobile ? (
-        <AccordionItem
-          isOpen={isAccordionOpen}
-          onToggle={toggleAccordion}
-          title={title}
-          icon={CaretDown}
-          iconSize={16}
-          className="[&_.accordion-toggler]:text-gray-400 [&_.accordion-toggler]:mb-2 sm:[&_.accordion-toggler]:mb-0 [&_.accordion-toggler]:text-4 [&_.accordion-toggler]:uppercase sm:[&_.accordion-toggler]:px-3.5 [&_.accordion-icon]:text-gray-700 mb-4 last:mb-0"
-        >
-          {RootItems}
-        </AccordionItem>
-      ) : (
-        <>
-          <button
-            type="button"
-            className="subnav-button gap-3 px-0 sm:px-3.5"
-            ref={setTriggerRef}
-          >
-            <Icon size={14} />
-            {label}
-          </button>
-          {visible && (
-            <PopoverBase
-              setTooltipRef={setTooltipRef}
-              tooltipProps={getTooltipProps}
-              withTooltipStyles={false}
-              cardProps={{
-                rounded: 's',
-                hasShadow: true,
-                className: 'py-6 px-2',
-              }}
-              classNames="w-full sm:max-w-[13.25rem] mr-2"
-            >
-              <>
-                <span className="text-4 text-gray-400 px-3.5 uppercase">
-                  {title}
-                </span>
-                {RootItems}
-              </>
-            </PopoverBase>
-          )}
-        </>
-      )}
-    </>
-  );
-}
 
 function Filter<TValue extends FilterValue>({
   items: rootItems,
@@ -242,7 +53,7 @@ function Filter<TValue extends FilterValue>({
   );
 
   const RootItems = rootItems.map(({ icon, items, label, name, title }) => (
-    <RootFilter
+    <FilterItem
       key={name.toString()}
       label={label}
       icon={icon}
@@ -255,6 +66,10 @@ function Filter<TValue extends FilterValue>({
     />
   ));
 
+  const filterCount = Object.values(value).reduce((acc, obj) => {
+    return acc + Object.values(obj).filter((item) => item === true).length;
+  }, 0);
+
   const [isSearchOpened, setIsSearchOpened] = useState(false);
 
   return (
@@ -265,6 +80,7 @@ function Filter<TValue extends FilterValue>({
           onClick={toggleModalOn}
           setTriggerRef={setTriggerRef}
           customLabel={buttonText}
+          numberSelectedFilters={isMobile ? filterCount : undefined}
         />
         {isMobile && (
           <Button
